Refactor CommandChannel processor
