+++ commands/addcommandtochannel.js
@@ -1,5 +1,5 @@
 const Permissions = require("../permissions.js");
-const PersistCommands = require("../persist/persist-commands.js");
+const CommandPermsProcessor = require("../processors/command-perms.js");
 
 exports.getPermissions = (client, message, args) => {
     return { User: Permissions.UserPermissions.ServerAdmin, Channel: Permissions.ChannelPermissions.All };
@@ -12,15 +12,5 @@ exports.run = async (client, message, args) => {
 	const channelId = message.channel.id;
 	const command = args[0];
 
-	console.log("addcommandtochannel invoked. command = " + command + " guildId = " + guildId + " channelId = " + channelId);
-
-	var commandData = await PersistCommands.readFromDisk();
-
-	var currentChannelCommands = PersistCommands.getCommandList(commandData, guildId, channelId);
-
-	if (currentChannelCommands.indexOf(command) === -1) {
-		currentChannelCommands.push(command);
-	}
-
-	await PersistCommands.writeToDisk(commandData);
+	return CommandPermsProcessor.addCommandToChannel(guildId, channelId, command);
 };
